Adding top holder class to the styles.scss file
Without this patch, the rules may "leak" across to other components.

This is a problem because if the effect exceeds its intended range,
the results cannot be predicted

This patch solves the problem by wrapping every single style rule
within the styles.scss in a top-level, which ensure that the rules
don't "leak" across to other components.

Issue-URL: https://app.asana.com/0/1201609890347122/1201780787043048/f

[x] The message explains WHY the change is needed.
[x] The message explains HOW the change solves the problem.
[x] The commit does not combine unrelated changes.
[x] The commit does not introduce any breakage.
[x] The commit does not introduce any new technical debt.
[] The commit updates any related documentation.
[] The commit includes automated testing.
[] The commit has been thoroughly tested.