dapp: fix import issues caused by MR!76
For some reason, MR!76 passed CI but failed post-merge into main.  So
address the resulting import issues, which seem to be related to a bug
with eslint-plugin-import where imports of type-only exports cannot be
mixed with normal imports:

  https://github.com/import-js/eslint-plugin-import/issues/2187

The workaround seems to be to separate them, and switch from the
BigNumberish type to the BigNumber class which seems more appropriate
anyway as tighter typing.