defining variables in variables.scss
Without this patch, we need to change the value of the variable in all
scss files where it is used.

This is a problem because if we want to change a property in the
project, we need to look for it in the whole codebase and change it.

This solves the problem because changing a variable in variables.scss
will change the value of the variable in all scss files.

Issue-URL: https://app.asana.com/0/1201609890347122/1201845837533349/f

[x] The message explains WHY the change is needed.
[x] The message explains HOW the change solves the problem.
[x] The commit does not combine unrelated changes.
[x] The commit does not introduce any breakage.
[x] The commit does not introduce any new technical debt.
[] The commit updates any related documentation.
[] The commit includes automated testing.
[] The commit has been thoroughly tested.