✨ adding notification for unstaking actions
Without this patch, the unstaking notifications are not shown.

This is a problem because it results in bad UX as a user is not
informed about the progress of the unstaking.

This patch solves the problem by adding success/error notifications
for unstaking actions. Moreover, this patch make notifications
not dismissible by click and, therefore, a user may click the
links or copy the text from success or error notifications. The fix
could be tested by staking and unstaking ZKP tokens in UI.

Issue-URL: https://app.asana.com/0/1201609890347122/1201754552015159/f

getEventFromReceipt function is extracted in the separate file