dapp: rename tokenUSDValue state variable to tokenBalanceUSD
tokenUSDValue was a very misleading name, because it was referring to
the USD value of the token *balance*, not the token price.  So rename
it and the corresponding setter to include the word "balance" in the
names.  Also drop the word "Value" because it's a variable so it's
self-evident that it's a value.