parsing of the errors returned from unsuccessful transactions
Without this patch, the error message is very long and unreadable.
It contains debug information that is not needed for the user.

This is a problem because the error message is displayed to the user
and the user may not be able to understand the error message.

This patch solves the problem by shortening an error message
extracting important information into the human readable string.
Parsing mechanism works on the assumption that an error object
returned after unsuccessful transaction contains error.error.message
and error.reason keys. Both of them or just error.error.message.
Otherwise, it falls back to the standard error.message.

Issue-URL: part of https://app.asana.com/0/1201609890347122/1201754552015159/f