Refactoring ErrorWithTx to MessageWithTx
Without this patch, in ErrorWithTx there is nothing in the body
of this function specific to errors. Also, there is no point
in including the full tx hash in the error message.

This is a problem because it results in bad UX as well as
misleads the developers with not very precise naming.

This patch solves the problem by renaming the component and
changing representation of link to the block explorer simply with
underlined "your transaction" link.

Issue-URL: part of https://app.asana.com/0/1201609890347122/1201754552015159/f