dapp: fix UnstakeTable updating infinitely
Without this patch, the UnstakeTable will be rendered infinitely

This is a problem because it is not acceptable to render component
when there is no need for that since the function is declared outside
of useEffect, we will have to put it in the dependency array of the
hook.  But if the function isn't wrapped in useCallback, it will
update on every re-render, and thus trigger the useEffect on every
re-render.  Which is not what we want here!

This patch solves the problem by wrapping the setTotalStaked function in useCallback hook

Issue-URL: https://app.asana.com/0/1201609890347122/1201784537900951/f

how to reproduce the problem being solved:

just remove the useCallback using, and the app will be rendered infinitely,
add console log message to see the problem
you can see the console log in the browser console