dapp: fix rounding bug with staking amount
Without this patch, the available unstaked token balance is stored and
displayed with only 2 decimal places, and can sometimes be rounded up.

This can result in an attempt to stake fractionally more than the
available balance, leading to a failed transaction and ETH wasted
when paying the gas fees.

This patch solves the problem by switching storage / handling of all
balances and USD prices to be BigNumbers, and only formatting in
human-readable form on display.

This allows us to safely revert commit "dapp: temporarily remove buggy
max button" (4c788c2ac16094092597913bed437fd72b67d1f3)

getRewardsBalance was unused, and getRewardsBalanceForCalculations was
an overly long name.  Since we are transitioning to handling all
arithmetic using BigNumber, and only converting to strings for UI
display purposes, there is no need for the "ForCalculations" suffix.
So remove getRewardsBalance and shorten
getRewardsBalanceForCalculations to getRewardsBalance.

Also deal with case where total staked is 0.  In development / staging
environments, calculating APY before anything is staked causes a
division by zero.  So add handling for that case.