dapp: updating tooltip with total staked tokens amount
Without this patch, we won't be able to see total amount of all $ZKP
tokens staked.

This is a problem as we are not showing this info anywhere on the
screen and users need to see amount of tokens staked, because that
affects the overall APY calculation.

This patch solves the problem by showing total amount of all $ZKP
tokens staked inside the tooltip on APY number element hover.