dapp: fix rounding down of currency values
When the browser locale uses something other than `.` as the decimal
separator, the regular expression for truncating the value string
didn't work, so the full precision was displayed.

So auto-detect the decimal separator, and also switch to using
navigator.language as the locale, since this seems more reliable based
on experiments.