dapp: fix accidental rounding to nearest integer
The previous enhancements to render currencies using the current
locale whilst rounding down had a couple of issues.  Firstly, when `.`
is the decimal separator, it was not correctly escaped in the regular
expression for truncation, so fix this using lodash's escapeRegExp().

Secondly, automatic detection of the separator was using the wrong
locale.  So add the locale as a parameter to toLocaleString().