Remove cause of horizontal scrollbar from unstake table
Without this patch, The unstake table is slightly too wide, causing a scrollbar.

This is a problem because it add unnecessary horizontal scrollbar to the stake table.

This patch solves the problem by  shrinking The column widths  and decreasing the fpnt size.

Issue-URL: https://app.asana.com/0/1201609890347122/1201854434453077/f

 Checklist:

  [ x ] The message explains WHY the change is needed.
  [ x ] The message explains HOW the change solves the problem.
  [ x ] The commit does not combine unrelated changes.
  [ x ] The commit does not introduce any breakage.
  [ x ] The commit does not introduce any new technical debt.
  [ ] The commit updates any related documentation.
  [ ] The commit includes automated testing.
  [ ] The commit has been thoroughly tested.

undo changes on unstakeTable design
Without this patch, the design is not compatible with figma

This is a problem because it break the UI

This patch solves the problem by redesigning the table as it is in figma
Issue-URL: https://app.clickup.com/t/2ky8z4e