dapp: round total amount staked in tooltip down to integer
Without this patch, the total amount staked which appears when
hovering on the tooltip icon in the APY box is displayed with
2 decimal places of precision.

This is a minor problem because it's more precision than needed,
creating a long number which is harder to read than necessary.

This patch solves the problem by rounding (down) to an integer.  It
required enhancements to utils.formatCurrency() and the corresponding
test suite, since it was not previously working correctly with the
`options` parameter set to `{decimals: 0}`.