using safeWindowOpen function instead of window.open in StackTab
Without this patch, we may face leak of the security of your website,
and allow other websites  gaining access to our page (through the browser session).

This is a problem because it reduce the security of the website.

This patch solves the problem by It prevents the opening page to gain any kind of access to the original page.
The safeWindowOpen function takes a URL as a parameter and opens a new window to that URL.

Issue-URL: https://app.clickup.com/t/1zc2c27