check that ETH balance is updated after staking and unstaking
Without this patch, ETH balance unupdated according to staking and unstaking process, which consume gas fees and assume
that the ETH balance will change according to that.

This is a problem because it is not possible to know the exact ETH balance  after staking and unstaking,unless we reload the page.

This patch solves the problem by fetching the ETH balance after each staking or unstaking process

Issue-URL: https://app.clickup.com/t/1zc37r2