using the generated avatar in the header
Without this patch, we have to use metamask svg icon in header.

This is a problem because it doesn't makes sense, and doesn't tell about which account is used.

This patch solves the problem by using the address as the hash to generate the avatar using react-jazzicon library.

-[x] The message explains WHY the change is needed.
-[x] The message explains HOW the change solves the problem.
-[x] The commit does not combine unrelated changes.
-[x] The commit does not introduce any breakage.
-[x] The commit does not introduce any new technical debt.
 [ ] The commit updates any related documentation.
 [ ] The commit includes automated testing.
 [ ] The commit has been thoroughly tested.

replacing  metamask icon with generated  from address

Without this patch, ...

This is a problem because ...

This patch solves the problem by ...

Issue-URL: ...