reorganise archives of Rinkeby staging artifacts and setup
Previously it was a bit confusing to store both actively
usable hardhat-deploy artifacts *and* inactive unusable
artifacts within the same contracts/deployments/ directory.

So create a new ARCHIVE subdirectory to clarify the difference
between those deployments which can be immediately used in hardhat,
and those which can't.

Also, since we now have a mumbai staging environment as well
as a rinkeby one, add an extra level of depth to the directory
tree to cater for both.

Clickup: #u7hu[completed]