setting up redux toolkit skeleton in the project
Without this patch, we won't be able to effectively handle data through the application.

This is a problem as the app grows our data and logic behind it grows and we need a state management library to handle all of our logical layer.

This patch solves the problem by adding redux stores with reducers and actions.

Issue-URL: https://app.clickup.com/t/21pear6