dapp: track unstaked balance in Redux instead of useState hooks
Without this patch, the unstaked balance tracked with useState hooks.

This is a problem because it's not easy to share useState state across multiple components, since you have to keep passing the getters and setters via props.

This patch solves the problem by changing the way the unstaked balance from using useState hooks to the Redux store.

Issue-URL: https://app.clickup.com/t/227j2bf