dapp: track current stake apy balance in Redux instead of useState hooks
Without this patch, the current stake apy is being tracked with useState hooks.

This is a problem because it's not easy to share useState state across multiple components, since you have to keep passing the getters and setters via props.

This patch solves the problem by changing the way the current stake apy is tracked from using useState hooks to the Redux store.

Issue-URL: https://app.clickup.com/t/22yej9m

Clickup: #22yej9m[in preview]