dapp: add token button - keep add token button visible
Without this patch, we will have issues with add token button - button will be hidden if added on one chain.

This is a problem because clicking on add token button we hide it without taking into consideration multichain environment.

This patch fixes this issue while keeping add token button visible all the time in the settings dropdown.