fixing division by zero in unstake table component
Without this patch, if the staking table is empty, it return an error because totalStaked is zero and we try to divide on it.

This is a problem because we cann't divide by zero

This patch solves the problem by checking if totalStaked is zero before dividing.
 if (totalStaked.gt(constants.Zero)) {...}

Issue-URL: https://app.clickup.com/t/1z4g4wf

Clickup: #1z4g4wf[completed]