contracts: fix 'currentBlock<startBlock+QUERY_BLOCKS' in vesting:list
Without this patch, the execution of the task fill skip
events if 'currentBlock < startBlock + QUERY_BLOCKS'
in the first iteration of the loop. For example,
startBlock=100
QUERY_BLOCKS=10
currentBlock=105

'while' statement will return false and the loop will not be
executed.

This patch solves the problem by changing 'while' condition.