AlertDialog component for staking warning
Without this patch, we can't warn stakers that if they stake now stake
will not start accruing rewards until the switchover is complete.

This is a problem because Staking is currently undergoing the
switchover described in DAO proposal #4.

This patch solves the problem by adding AlertDialog to warn stakers
about this point before staking.

Issue-URL: https://app.clickup.com/t/239y2qr

Clickup: #239y2qr[in review]