dapp: fix regression with total rewards
The previous "fix" in 3e65165e was intended to avoid showing wrong
Polygon reward balances when switching to an unsupported network.  But
it called toNumber() on rewardsBalance which resulted in an overflow.

Also, the balance should be reset whenever on an unsupported network,
so move the reset higher in the call stack.