contracts: fix bugs in paginator
It was calculating the query end block wrong, so that if there was
only one chunk, it would always be the same size and potentially
extend past the end block.

There was an off-by-one error calculating the number of chunks, so fix
that.

Finally, it was using a range including the requested starting block
but excluding the requested end block, which was counter-intuitive in
this scenario.  So include the end block.