contracts: rerequire function to re-import modules in case of update
Without this patch, we had to restart node console if the code of some
functions were updated.

This is a problem because it slows down the development process.

This patch solves the problem by adding the wrapper around require
function that deletes the module from the cache and reimports it.
Example:
```
$ HARDHAT_NO_MNEMONIC=yes hardhat console --network localhost
Welcome to Node.js v14.18.3.
> r = require('./lib/rerequire');
{ rerequire: [Function: rerequire] }
> t = r.rerequire('./lib/test')
{ test: [Function: test] }
> t.test()
{ a: 2 }
> t = r.rerequire('./lib/test')
{ test: [Function: test] }
> t.test()
{ a: 1 }
```