dapp: left align and move max button on top of input field
Without this patch, we won't have user intuitive approach on input field with max button and right alignment.

This is a problem because we need to reposition some elements around the input field for better UX.

This patch fixes this issue while left aligning content inside of input field and moving max button on top of the field.