dapp: eslint to add a newline after import
Without this patch, the import statements in the header
of the file were not separated by the empty line.

This is a problem because it results in bad readability
of the code and break styling conventions.

This patch solves the problem by adding a rule of "newline
after import" to the eslint config. CI pipeline picks up
the rule as well and shoots an error if it is not obeyed.