try to fix ClickUp / GitLab integration
Without this patch, we have no "magic" reference to a ClickUp task
which ClickUp will automatically detect and create a link from the
task to the merge request.

This is a problem because it means ClickUp tasks need to be
manually linked to merge requests.

This patch attempts to solve the problem by adding the magic line to
the merge request template.  However the documentation is not clear
on whether this will work, so let's try it as an experiment.