Initial import of panther-pool-client
Without this patch, we would have logic scattered across repositories.

This is a problem because it makes contributions more difficult for review and QA. Also requires more effort to develop and set local env etc.

This patch solves the problem by importing the elements we need from previous panther-pool-client to panther-core.

Issue-URL: https://app.clickup.com/t/294rv1u

Clickup: #294rv1u

This commit only imports files as is and restructuring will follow in next.

[ ] The message explains WHY the change is needed.
[ ] The message explains HOW the change solves the problem.
[ ] The commit does not combine unrelated changes.
[ ] The commit does not introduce any breakage.
[ ] The commit does not introduce any new technical debt.
[ ] The commit updates any related documentation.
[ ] The commit includes automated testing.
[ ] The commit has been thoroughly tested.