dapp: adding back advanced staking option and dropdown
Without this patch, we won't be having 2 options of stake type to choose on.

This is a problem because we need to have both normal and advanced staking methods which user can choose.

This patch fixes this issue while reverting back rendering of staking method dropdown and adding logic to pass on selected staking type to the stake method call. In detail, the changes are the following:

- staking method dropdown - logic for getting the stakeType and parsing it to the stake method
- parsing update stake method type to the stake method contract call
- parsing update stake method type - adding missing useEffect dependency
- fixing dropdown position and background color
- fixing dispatch instance