dapp: source maps to show original TypeScript code
Without this patch, when we follow webpack://@zkp-staking/dapp  URLs from the browser's dev console to view source corresponding to frames of a stack trace, we get transpiled Javascript code rather than the original TypeScript we've written.

This is a problem because it makes debugging quite a bit harder.

This patch solves the problem by fixing generation of the source map.

Issue-URL: https://app.clickup.com/t/2phmgaq