add a feature flag for advanced staking
Without this patch, we have no way of simultaneously supporting
classic staking and advanced staking.

This is a problem because we don't intend to initially support
advanced staking on Ethereum, and also we can't afford the risk and
cost of maintaining classic and advanced staking branches in parallel.

This patch solves the problem by adding a feature flag to enable
advanced staking per chain.  This will allow the staking type dropdown
and other advanced staking behaviour to only appear on chains where
advanced staking is supported.