dapp: fixing type generation for packages imported from pool-client code
Without this patch, build is failing in dapp with the message of missing types.

This is a problem because we cannot build application.

This patch solves the problem by importing gen-types.sh script from preZKP to dapp which builds those missing types.

This commit contains change in tsconfig file where we add "types/*" in the path array and while types are actually found in "src/types" this particular path doesn't fix the broken build. It might be because of webpack build process and the way types are stored after webpack build.
It also contains change in gitlab CI where instead of just building contract types we also build dapp types because not doing this make CI fails and is necessary step to build dapp.