dapp: derive seed from signed message
Without this patch, we would not be able to derive seed from signed message passed onto crypto layer from react app.

This is a problem because we need seeds in order to derive root keypairs.

This patch solves the problem by implementing `deriveSignatureFromSignedMessage` and `deriveSeedFromSignedMessage` functions and adding tests for those.

Issue-URL: https://app.clickup.com/t/2td376t