dapp: deterministic generation of the key pair from the signature
Without this patch, we cannot generate keypairs in a deterministic way.

This is a problem because we need to generate and re-generate the keys
in case of the loss of the derived keys.

This patch solves the problem by generating the keypair from the signature
in a deterministic way. Added tests for the new code.

Issue-URL: https://app.clickup.com/t/2td377z