dapp: tests for generation of the random keypairs
Without this patch, generation of the random keypairs
is not covered by the tests.

This is a problem because we need to be sure that the keys
are withing SNARK_FIELD_SIZE and non-deterministic.

This patch solves the problem by writing tests that ensure
that the keys are within the range of SNARK_FIELD_SIZE and
are deterministic.

Issue-URL: https://app.clickup.com/t/2td88x9