dapp: adding keypair derivation method to services
Without this patch, we cannot call keypair derivation from react layer in dapp.

This is a problem because we need to generate root keypairs by clicking on advanced stake button in order to create staking input parameters.

This patch solves the problem by implementing deriveKeypairFromSignature() in services layer and adding related tests to it.

Issue-URL: https://app.clickup.com/t/2td9g2a