dapp: multiplication of the private keys
Without this patch, we cannot generate products of the private keys

This is a problem because we need to generate a Spender keypair which
is the product of 1) Random secret scalar and 2) Recipient root spending
private key. For details, check MASP Rosetta stone and Shielded pool
design documents for additional information.

This patch solves the problem by implementing the multiplication function.
This function is covered by tests to ensure that multiplication is
commutative.

Issue-URL: https://app.clickup.com/t/2td9nq9