dapp: generation of a Spender keypair
Without this patch, we cannot generate a Spender keypair.

This is a problem because a Spender keypair is needed as part
of the UTXO commitment to ensure that only the recipient can
spend the UTXO. For details, check MASP Rosetta stone and
Shielded pool design documents for additional information.

This patch solves the problem by implementing generateSpenderKeypair
function in the service layer.

Issue-URL: https://app.clickup.com/t/2td9jpm