dapp: ECDH encryption/decription in service layer
Without this patch, we would not be able to call encrypt / decrypt message functions from react UI layer.

This is a problem because we need to encrypt (and later decrypt) advanced staking parameters for zZKP and PRP staking.

This patch solves the problem by implementing encryptCommitment and decryptCommitment functions in service layer of dapp.

Issue-URL: https://app.clickup.com/t/2td9mkr