dapp: show staking method dropdown only if the chainHasAdvancedStaking is returning true
Without this patch, we won't be having an ability to show/hide staking method if the flag is set to true/false.

This is a problem because we need to have control of this feature and turn it on/off as needed.

This patch fixes this issue while showing/hiding the staking method dropdown based on the flag value from env variables