dapp: fix breakage from recent merges of encryption code
The recent MRs broke on merge, which points to at least one
flaw in our CI.  Fix the tests which broke for reasons relating
to the switch to AES-256-CBC.  In particular the ciphertext
is not in the field element format the tests expected, so fix
those expectations.

Also don't catch exceptions at a low level from failed decryptions;
that can be done at a higher level.

Fixes #8.