dapp: adding support to UI Stake button for advanced staking
Without this patch, we cannot stake using advanced stake type
from UI.

This is a problem because we need to enable users to stake in our app.

This patch solves the problem by adding conditional logic into stake()
service function, encripting ethmeral keys in messages and preparing
the data as byte code.

Issue-URL: https://app.clickup.com/t/2bmu679