debug: adding debugging information for the test that randomly fails
Without this patch, it is very difficult to catch a randomly failing
test of aes256 encryption.

This is a problem because sometimes for no obvious reason the test fails.
This test checks that decryption should throw an error if it is attempted
with an incorrect key.

This patch solves the problem by printing decrypted text that shows that
it is not matching with the original information. This is to ensure that
it is not possible to decrypt the message with an incorrect key. I've
run the test locally 1000 times and it failed 5 times. In these failed
tests, I confirmed using this debugging output that decrypted text is
incorrect.

Most likely, It is an issue with the underlying crypto library we are using.