dapp: disable staking operation since classic staking closed
Without this patch, testing staking on 27th April on Polygon yields a
smart contract error "Transaction is not allowed anymore". This
happened because classical staking is over but advanced staking is not
started yet.  The UI did not show any errors or notifications about
the problem and I was not sure what is happening with the stake
action.

This patch solves the problem by disabling the staking input field and
button, and updating the text to explain.

Issue-URL: https://app.clickup.com/t/2uffapn