contracts: hardhat task to add advanced terms in the local environment
Without this patch, the forked staking contract does not have terms for advanced stakes.

This is a problem because we cannot execute advanced stakes in the development environment.

This patch solves the problem by adding a task that Adds terms to the staking contract for advanced staking in a local hardhat. The task was written for the local env. because we would like to test advanced stakes in a forked environment. It is problematic because we cannot write a task that simply adds terms. After all, these terms need to be added by the owner of the smart contract. Therefore, this task has impersonating and top-up balance functions. Also, we could not use already written methods for impersonation and increasing balance written in 'contracts/lib/hardhat.ts', because this file imports hardhat (hh), and hh tasks cannot use files which have hh import: https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/268

Issue-URL: https://app.clickup.com/t/2bmu679