dapp: adding service function to calculate rewards for a stake
Without this patch, we are not able to calculate the rewards for an
advanced stake.

This is a problem because we need to display rewards in UI before
the advanced stake, as well as we need to show rewards for the
existing stakes.

This patch solves the problem by implementing functions that calculate
zZKP and PRP rewards for an advanced stake, as well as fetches rewards
from the contract for classic stakes. It also adds tests and fixes CI,
which were failing due to the addition of the new Rewards type in StakeRow.

https://app.clickup.com/t/2c63yd9