dapp: new UI for zAssets components
Without this patch, we won't be able to render all zAssets on the ZAssets page with the new refreshed UI.

This is a problem because we need to align the UI fo zAsset components according to the new Figma design.

This patch fixes this issue while adding zAssets subcomponents and splitting them into smaller parts. Adding correspnding routes.