contracts: refactoring of the staking:list task
Without this patch, staking:list task can only fetch StakeCreated events. Also,
events are always saived in chunks inspite of --out flag.

This is a problem because we need to fetch StakeClaimed events from staking contract.

This patch solves the problem by additng additional task argument to the staking:list
task 'filter' which specifies name of the event. To fix always saived events in
chunks, additional check is added if the task argument --out is provided.