dapp: renaming and refactoring of txError function
Without this patch, txError function has a misleading name and is used only to display
transaction errors.

This patch solves the problem by renaming the function to notifyError.
Then, the addition of the title argument to the function signature made the function
more generic and able to notify any type of error in UI. Therefore, this function is
used also to handle other errors during staking and unstaking such as those returned
when an event was not found in the receipt.