dapp: support for new StakeRewardController2 contract
A newly deployed StakeRewardController2 contract that is replacing RewardMaster contract has a different name.
Without this patch, Panther dApp uses variable names related to RewardMaster. This is a problem because variable
and function names mismatch the contract name.

This patch solves the problem by renaming existing variables and functions in dApp that were used for RewardMaster
contract with new names related to StakeRewardController2 contract. Alongside, contract ABI is updated, and the
env.example file to use the new contract.

Issue-URL: https://app.clickup.com/t/29j5a3j