dapp: new stake/unstake ui tabs
Without this patch, our stake/unstake tabs wouldn't look
according to the new figma designs.

This is a problem because we need to have updated look/layout
of new ui for stake/unstake tabs.

This patch fixes this issue while new components/subcomponents
for the ui on stake/unstake tabs.