dapp: updated message for keypair derevation
Without this patch, the message that user signs to generate
the keypair doee not reflect the fact that is free of cost.

This is a problem because some users may have a perception
that they are paying for the keypair generation.

This patch solves the problem by adding a message to the
keypair generation dialog clearly stating that signing will
cost nothing.. The message is:

Greetings from Panther Protocol!
Sign this message in order to obtain the keys to your Panther wallet.
This signature will not cost you any fees.
Keypair version: 1

Issue-URL: https://app.clickup.com/t/26gtpr4