dapp: better RPC error handling
Without this patch, some of the errors were masked or
returned incomplete information.

This patch solves the problem by parsing error message with regexp
and returning the correct trancated one. In the other, this patch better parses
the error message when where is concise error representation in "reason" key.
Moreover, this patch adds specific testing cases.