dapp: refactoring reward.ts
Without this patch, 1) thrown error during rewards calculation was shadowed and there
was no sign if something went wrong. 2) thrown error did not allow calculation for the
stakes if the staking program was not started. That is the problem for developers in
development mode. 3) TokenID enum was not properly used and was replaced by a regular
string, e.g., 'ZKP'.

This patch solves problem 1) by catching the error and printing it in the console and
re-throwing it again. 2) Setting ADVANCED_STAKING_T_START and ADVANCED_STAKING_T_END
as environmental variables (that required additional mocking of env variables in tests).
3) Strings are replaced by TokenID enum.