dapp: gracefully handle trying to add Hardhat network
MetaMask has a bug which doesn't allow adding http://...
URLs via wallet_addEthereumChain, which means that in
particualr it can't add http://localhost:8545.

So watch out for this error, and handle it more gracefully.
Additionally fix a bug in how the failure notification is generated,
by passing the message string rather than the full error object.