dapp: fix warning: validateDOMNesting 'div cannot appear as a descendant of p'
Without this patch, console will still show this warning.

This is a problem because it broke HTML, but more importantly it
spams the console and makes it harder to spot other errors.

This patch solves the problem by replacing paragraph (where div is a
descendant of) with div (Box in material-ui).

Issue-URL: https://app.clickup.com/t/2aewqjn