dapp: fetching Staking Terms from the chain and saving in redux store
Without this patch, there is no way to fetch staking terms on the specific chain. This
is the problem because our UI conditionally shows different components based on the
staking terms.

This patch solves the problem by fetching the terms from the chain and setting
saving them into the redux store. Then, in different components (where needed) the
values are retrieved via a created selectors.

Issue-URL: https://app.clickup.com/t/2xgnqnb