dapp: fix dispatch of redux-thunk masking errors
Without this patch, during dispatching of the redux action, if any error
occurs, the redux-thunk will mask the error and will not show any sign of it.

This is a problem because the error is not logged in the console and the user
or developer will not be able to see if something is going wrong.

This patch solves the problem by creating a new dispatch function dispatch
(thunk, ...args) that will log the error in the console and then call the
original dispatch function. Please, be aware that the signature of the
dispatch function is changed and now it accepts a thunk as the first argument
and optional arguments for thunk call: thunk(...args).