dapp: calculated rewards in redux store
Without this patch, our zZkp and prp rewards are calculated in the comopnent state.

This is a problem because setZZkp and setPrp are called by fetchStakingRewards, which
is called by setStakingAmount, which is defined in StakeTab, but passed to
StakingInput, and the corresponding getters zZkp and prp are used in
AdvancedStakingRewards.  So the app had the useState definitions in one component, the
setters used in a second, and the getters used in a third.

This patch fixes this issue while adding calculatedRewards slice which are now pulled
by selectors in UI. Additionally, this patch creates shared StakeRewards
and StakesRewardsState types used across different redux slices.