dapp: fix error when safeDispatching action which is not thunk
Without this patch, when you dispatch an action, it will be dispatched by the new
safeDispatch hook. This hook assumes that all actions are thunks. In the case when
action is not thunk, the safeDispatch hook throws an error as a returned promise has
no unwrap() method.

This patch solves the problem by checking promises returned from dispatching thunks
that they are actually can be unwrapped.