dapp: refactoring of color #ffdfbd in CSS
Color #ffdfbd is one of the main colors of the theme. However, in many places it is
used not as a global variable but as a hardcoded hex values of the color.

This patch solves the problem by setting the name of the variable and replacing hex
values with global variable name.