dapp: faucet page for sending test tokens on Mumbai testnet
Without this patch, it is problematic to send a user test tokens and share private keys
with the community. As shown in our recent experience, even testing tokens can be stolen
even though they have no value.

This is a problem because we need to give a user frictionless experience of testing our
protocol.

This patch solves the problem by adding a faucet page where users can request tokens
directly in their Metamask account.