dapp: Fix max button styling
Without this patch, the max button in the staking tab is not styled correctly.

This is a problem because the max button doesn't looks like a button, as it looks in Figma.

This patch solves the problem by changing the max button style in styles.scss file

Issue-URL: https://app.clickup.com/t/2yda7p0

Checklist to make sure this is a high quality commit:

- [x] The commit messages all explain WHY the changes are needed.
- [x] The commit messages all explain HOW the changes solve the problem.
- [x] The commits are all strongly related.
- [x] None of the commits introduce any breakage.
- [x] None of the commits introduce any new technical debt.
- [ ] The commits update any related documentation.
- [ ] The commits include automated testing.
- [x] The commits have been thoroughly tested.