dapp: Implement zZKP redemption on zAsset page
Without this patch, we can't redeem zZKP in zAsset page.

This is a problem because a user can't redeem zZKP.

This patch solves the problem by adding **redeem zZKP** button for each row in the private zAsset table as well as
introducing an accordion-style detailed dropdown. Each row of the dropdown represents the unspent zZKP UTXO of the user.

Issue-URL: https://app.clickup.com/t/2y7t27u