faucet: balance refresh
Without this patch, total balance is not refreshed on any action execution.

This is a problem because we need to have balance refreshing on every action.

This patch fixes this issue while adding a part of the code which calls dispatch of zkp token balance fetching at the correct actions.