refactor by extracting type commitment into @panther-core/crypto
Without this patch, the commitment type is used only in one hardhat task
of the @panther-core/contracts workspace. It is ok until we introduce a
new script in @panther-core/crypto workspace to build a Merkle tree that
uses the same type. To avoid the creation of the circular references of
the commitment type, we extract it into low-level @panther-core/crypto
workspace. Also, NewCommitment is renamed to NewCommitmentLog to be more
descriptive as commitments are inside this structure.