crypto: generate Merkle Tree based on commitments from NewCommitments events
Just for advanced staking, we need to periodically build the MASP v0
Merkle tree and make it available for download by the dApp, for users to
be able to activate the early zZKP redemption.

This patch solves the problem by creating a script that generates and
saves Merkle tree in from the JSON file of commitments. Script reads
NewCommitmentEvent[] in the provided JSON file and form tree. Also,
added tests of the tree root generated by smart contract and
documentation.

Arguments:

    1. --file or -f path to the JSON file with an array of NewCommitmentLog
    2. --path or -p path to the folder where the tree will be saved in JSON format

Output:

    - JSON file of the tree with the commitments in the specified folder:
    identities-tree-0.json

Example:

    yarn ts-node scripts/build-commitments-tree.ts \
     -f tmp/events/20220713-0x6B982bE424aEe27e0251eE81dA024Eb25C83C497-mumbai-full.json \
     -p ../dapp/public/

Issue-URL: https://app.clickup.com/t/2au3w9z